First draft section 4
Could use screenshots of in game GUI for the terms as well as the save file rockets. Lots of free time during/after this to explore the game/orbits. More complex orbital mechanics are sadly out of scope; possibly a small section about making better rockets if there's time remaining.